Room Type Converters

+++ app/src/main/java/com/example/weatherapp/models/Main.kt
@@ -6,8 +6,8 @@ import com.google.gson.annotations.SerializedName
 data class Main(
     @SerializedName("feels_like")
     val feelsLike: Double,
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val temp: Double,
     @SerializedName("temp_max")
     val tempMax: Double,
