+++ app/src/main/java/com/example/weatherapp/models/Wind.kt
@@ -4,6 +4,6 @@ package com.example.weatherapp.models
 import com.google.gson.annotations.SerializedName
 
 data class Wind(
-    val deg: Int,
+    val deg: Double,
     val speed: Double
 )
